backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
celery>=5.3.6
redis>=5.0.1
openai>=1.10.0
google-generativeai>=0.3.2
python-dotenv>=1.0.1
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
scipy>=1.11.0
